+++ ui/proxima-consulting-hub/src/services/base-api-gateway.ts
